Move common hover handling into helper
